requirements.txt
python-telegram-bot==13.12
pandas==1.3.3
configparser==5.0.2
